scripts/requirements.txt
telethon>=1.34
schedule>=1.2
Pillow>=10.0
# Optional for generate_sitemap push:
google-auth>=2.0
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.2
